+++ rpnpy/functions.py
@@ -303,6 +303,29 @@ def list_(calc, modifiers, count):
 list_.names = ('list',)
 
 
+def store(calc, modifiers, count):
+    """Store some stack items into a variable.
+
+    @param calc: A C{Calculator} instance.
+    @param modifiers: A C{Modifiers} instance.
+    @param count: An C{int} count of the number of arguments to pass.
+    """
+    variable, args = calc.findStringAndArgs('store', modifiers, count)
+    if len(args) == 1:
+        # Only one argument from the stack, so we'll set the variable to
+        # have that value.
+        value = args[0]
+    else:
+        value = args
+
+    calc.setVariable(variable, value)
+    calc._finalize(None, nPop=len(args) + 1, modifiers=modifiers, noValue=True)
+    return calc.NO_VALUE
+
+
+store.names = ('store',)
+
+
 def map_(calc, modifiers, count):
     """Map a function over some arguments.
 
@@ -343,6 +366,7 @@ FUNCTIONS = (
     reduce,
     reverse,
     stack,
+    store,
     swap,
     undo,
     variables,